chore: Made test for sql serve schema issue sql-server specific.  I suspect the failing tests indicate the bug is still present on other DBMS's, but I don't know enough to go chasing that right now.
