+++ grate.unittests/SqlServerCaseSensitive/Running_MigrationScripts/Versioning_The_Database.cs
@@ -6,7 +6,7 @@ namespace grate.unittests.SqlServerCaseSensitive.Running_MigrationScripts
     [TestFixture]
     [Category("SqlServerCaseSensitive")]
     // ReSharper disable once InconsistentNaming
-    public class Versioning_The_Database : Generic.Running_MigrationScripts.Versioning_The_Database
+    public class Versioning_The_Database : grate.unittests.SqlServer.Running_MigrationScripts.Versioning_The_Database
     {
         protected override IGrateTestContext Context => GrateTestContext.SqlServerCaseSensitive;
     }
